Update mully_node.R
Due to a change in R 4.1.3 the `is.na()` function provides a vector of Booleans (for each column) instead of a single TRUE/FALSE value. Then using the `==` operator results in an error where the `length(...) = ... > 1' in coercion to 'logical(1)'` (see also: [stackoverflow.com](https://stackoverflow.com/questions/72848442/r-warning-lengthx-2-1-in-coercion-to-logical1)). The easy way out of this is to add the `all()` function, merging the Boolean vector into a single Boolean.